Submitting HW--citipix
Not so sure of some things in this hw, and it took me a long time to
figure out very basic things here. I was confused re. attributes, which
I learned (or re-learned) describe only HTML elements and jquery
variables—right???? I wasn’t sure how to incorporate the addAttribute
command for this reason. IDK. I also had an issue with having “New York
City” work as a string. Maybe a spaces thing, idk? Couldn’t find an
answer after googling for a long time. And finally, I couldn’t figure
out how to have a user continuing to submit city names after entering
the sydney values, which are written at the end of my function. I am
really hoping to get an answer for that. Feel  6/10 in terms of mastery
of this stuff, still a major struggle to do the most basic coding in
jquery for me! I probably have to review all of the material, re-read
notes and practice the vocab, which will take a long time and is
probably why i haven’t done it yet this week while i muddled through
this homework! ;-)